Revamp statistics collection
Our new approach is to record simulation events that allow us to replay
the simulation afterwards and obtain any statistics we need without
bloating the simulation code. A nice side-effect is that we can use the
same replay mechanism to validate that the simulation behaves as we
expect (this is partially done in the current SimulationStepper, but
needs to be extended to cover all properties we are interested in).